requirements.txt
fastapi
uvicorn
Pillow
piexif
#python-multipart  # for file uploads
